# openapi/integration-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сервис ремонта оборудования — интеграционный API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Интеграционный API, версия 1"
    }
  ],
  "paths": {
...
    "/service-bindings": {
      "get": {
        "tags": [
          "service-bindings"
        ],
        "summary": "List Bindings",
        "description": "Привязки к этому исполнителю, включая ожидающие его ответа (`status=pending`).\n\nТребует scope `service_bindings:read`.",
        "operationId": "list_bindings_service_bindings_get",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProviderBindingView_"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не пройдена аутентификация.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или недоступен вызывающему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния: версия, идемпотентность или недопустимый переход.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка проверки полей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен, повторите позже.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка. Формат одинаковый для всех методов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-scopes": [
          "service_bindings:read"
        ]
      }
    },
...
    "/service-binding-invitations": {
      "post": {
        "tags": [
          "service-bindings"
        ],
        "summary": "Create Invitation",
        "description": "Только допущенный (проверенный) исполнитель. Токен и ссылки приглашения показываются один раз: повтор того же `Idempotency-Key` их не выдаёт.\n\nТребует scope `service_bindings:write`.",
        "operationId": "create_invitation_service_binding_invitations_post",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindingInvitationCreateBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BindingInvitationIssuedView"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не пройдена аутентификация.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или недоступен вызывающему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния: версия, идемпотентность или недопустимый переход.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка проверки полей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен, повторите позже.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка. Формат одинаковый для всех методов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-scopes": [
          "service_bindings:write"
        ]
      }
    },
    "/service-binding-invitations/{invitation_id}/revoke": {
      "post": {
        "tags": [
          "service-bindings"
        ],
        "summary": "Revoke Invitation",
        "description": "Отозвать своё ещё не использованное приглашение.\n\nТребует scope `service_bindings:write`.",
        "operationId": "revoke_invitation_service_binding_invitations__invitation_id__revoke_post",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invitation Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BindingInvitationView"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не пройдена аутентификация.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или недоступен вызывающему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния: версия, идемпотентность или недопустимый переход.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка проверки полей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен, повторите позже.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка. Формат одинаковый для всех методов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-scopes": [
          "service_bindings:write"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "BindingInvitationCreateBody": {
        "properties": {
          "customer_inn": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Customer Inn"
          },
          "contract_number": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Contract Number"
          },
          "basis": {
            "type": "string",
            "title": "Basis",
            "default": "service_contract"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "equipment_items": {
            "items": {
              "$ref": "#/components/schemas/BindingInvitationItemBody"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Equipment Items"
          },
          "guarantor_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "manufacturer",
                  "seller",
                  "service_org"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Guarantor Kind"
          },
          "guarantor_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Guarantor Name"
          }
        },
        "type": "object",
        "required": [
          "customer_inn",
          "contract_number"
        ],
        "title": "BindingInvitationCreateBody"
      },
      "BindingInvitationIssuedView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "customer_organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Organization Id"
          },
          "contract_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Number"
          },
          "basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Basis"
          },
          "equipment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Equipment Ids"
          },
          "equipment_descriptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Equipment Descriptions"
          },
          "equipment_items": {
            "items": {
              "$ref": "#/components/schemas/BindingInvitationItemView"
            },
            "type": "array",
            "title": "Equipment Items"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "token_prefix": {
            "type": "string",
            "title": "Token Prefix"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "declined_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declined At"
          },
          "decline_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decline Reason"
          },
          "guarantor_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guarantor Kind"
          },
          "guarantor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guarantor Name"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "webapp_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webapp Link"
          },
          "bot_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bot Link"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state",
          "customer_organization_id",
          "contract_number",
          "basis",
          "equipment_ids",
          "equipment_descriptions",
          "equipment_items",
          "expires_at",
          "token_prefix",
          "created_at",
          "token",
          "webapp_link",
          "bot_link"
        ],
        "title": "BindingInvitationIssuedView"
      },
      "BindingInvitationItemBody": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Description"
          },
          "serial_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "BindingInvitationItemBody"
      },
      "BindingInvitationItemView": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "serial_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "index",
          "description",
          "serial_number",
          "model"
        ],
        "title": "BindingInvitationItemView"
      },
      "BindingInvitationView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "customer_organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Organization Id"
          },
          "contract_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Number"
          },
          "basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Basis"
          },
          "equipment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Equipment Ids"
          },
          "equipment_descriptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Equipment Descriptions"
          },
          "equipment_items": {
            "items": {
              "$ref": "#/components/schemas/BindingInvitationItemView"
            },
            "type": "array",
            "title": "Equipment Items"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "token_prefix": {
            "type": "string",
            "title": "Token Prefix"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "declined_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declined At"
          },
          "decline_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decline Reason"
          },
          "guarantor_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guarantor Kind"
          },
          "guarantor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guarantor Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state",
          "customer_organization_id",
          "contract_number",
          "basis",
          "equipment_ids",
          "equipment_descriptions",
          "equipment_items",
          "expires_at",
          "token_prefix",
          "created_at"
        ],
        "title": "BindingInvitationView"
      },
      "BindingPartyView": {
        "properties": {
          "organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_platform_member": {
            "type": "boolean",
            "title": "Is Platform Member"
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "name",
          "is_platform_member"
        ],
        "title": "BindingPartyView"
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Машиночитаемый код ошибки, например NOT_FOUND."
              },
              "message": {
                "type": "string",
                "description": "Сообщение на русском для лога/UI."
              },
              "request_id": {
                "type": "string",
                "description": "Идентификатор запроса — тот же, что в заголовке X-Request-ID."
              },
              "details": {
                "type": "object",
                "description": "Дополнительные поля ошибки, зависят от кода.",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "Page_ProviderBindingView_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProviderBindingView"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[ProviderBindingView]"
      },
...
      "ProviderBindingView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_explanation": {
            "type": "string",
            "title": "Status Explanation"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "basis": {
            "type": "string",
            "title": "Basis"
          },
          "contract_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Number"
          },
          "customer": {
            "$ref": "#/components/schemas/BindingPartyView"
          },
          "equipment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BindingEquipmentView"
              },
              {
                "type": "null"
              }
            ]
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "status_explanation",
          "status_reason",
          "basis",
          "contract_number",
          "customer",
          "equipment",
          "valid_from",
          "valid_until",
          "created_at"
        ],
        "title": "ProviderBindingView"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Ключ интеграции сервисной компании: `Authorization: Bearer <ключ>`.",
        "scheme": "bearer",
        "bearerFormat": "rk_<env>_<prefix>_<secret>"
      }
    }
  }
}